moving location of config to a place that makes more sense
