+++ modules/core/client/components/Map/constants.js
@@ -1,7 +1,6 @@
 import { getMapBoxToken } from '../../utils/map';
 import osmStyle from './osm.json';
 // don't know where to put this
-export const brandName = 'Bicycle Touring Community';
 export const MAP_STYLE_MAPBOX_OUTDOORS = 'mapbox://styles/mapbox/outdoors-v11';
 export const MAP_STYLE_MAPBOX_SATELLITE =
   'mapbox://styles/mapbox/satellite-streets-v11';
